Added new version of OSM Contributor.
The sample menu now launches the new OSM contributor.

Also changed OpenStreetMapActivity so that it does not deal with the LocationManager. As an abstract class, if a user requires that functionality it can be provided in the sub-class.

Still need to create a new LinearOverlay, that can be used in the GPSCaptureActivity to draw lines of where the user has been.

Also, GPX files are only saved on the SD Card - they are not uploaded yet.